credit-intelligence-platform/requirements.txt
# Data
pandas==2.2.2
pyarrow==16.0.0
fredapi==0.5.2
requests==2.32.3
kaggle==1.6.17

# Features
scikit-learn==1.4.2
numpy==1.26.4
scipy==1.13.0

# Models
xgboost==2.0.3
lightgbm==4.3.0
shap==0.45.1
mlflow==2.13.0
optuna==3.6.1

# Serving
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
redis==5.0.4
psycopg2-binary==2.9.9
kafka-python==2.0.2

# LLM
openai==1.30.1
httpx==0.27.0

# Monitoring
evidently==0.4.30
prometheus-client==0.20.0
apscheduler==3.10.4

# Evaluation
matplotlib==3.8.4
seaborn==0.13.2
fairlearn==0.10.0
calibration-belt==0.1.4

# Testing & Load
pytest==8.2.0
locust==2.28.0
